# openapi/v1/openapi.yaml
openapi: 3.1.0
info:
  title: ScamNemesis API
  description: |
    REST API for the ScamNemesis fraud reporting and tracking platform.

    ## Authentication
    The API supports three authentication methods:
    - **Bearer Token**: JWT access token in Authorization header
    - **HttpOnly Cookie**: Automatic for browser clients
    - **API Key**: X-API-Key header for programmatic access

    ## Rate Limiting
    - Standard users: 100 requests/hour per endpoint
    - Admin users: 1000 requests/hour (10x multiplier)
    - Auth endpoints: 20 requests/15min (brute force protection)

    ## Error Handling
    All errors follow RFC 7807 Problem Details format with additional fields.
  version: 1.0.0
  contact:
    name: ScamNemesis Support
    url: https://scamnemesis.com/contact
    email: support@scamnemesis.com
  license:
    name: Proprietary
    url: https://scamnemesis.com/terms
servers:
  - url: https://scamnemesis.com/api/v1
    description: Production server
  - url: https://staging.scamnemesis.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange refresh token for new access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
              required: [refresh_token]
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      description: Send password reset email
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required: [email]
      responses:
        '200':
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/2fa/verify-login:
    post:
      tags: [2FA]
      summary: Verify 2FA during login
      description: Complete login by verifying TOTP code
      operationId: verify2FALogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                temp_token:
                  type: string
                code:
                  type: string
                  pattern: '^\d{6}$'
              required: [temp_token, code]
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid code
        '401':
          description: Invalid or expired temp token
...
  /search:
    get:
      tags: [Search]
      summary: Search fraud reports
      description: |
        Search reports using multiple modes:
        - **auto**: Automatically detect search type and try exact then fuzzy
        - **exact**: Exact match on normalized fields (email, phone, IBAN, wallet)
        - **fuzzy**: Trigram-based fuzzy name matching
        - **semantic**: Vector embedding similarity search

        Results are masked based on user role.
      operationId: searchReports
      security:
        - {}
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          description: Search query
          schema:
            type: string
            maxLength: 500
        - name: mode
          in: query
          description: Search mode
          schema:
            type: string
            enum: [auto, exact, fuzzy, semantic]
            default: auto
        - name: fields
          in: query
          description: Comma-separated fields to search
          schema:
            type: string
        - $ref: '#/components/parameters/fraudType'
        - $ref: '#/components/parameters/country'
        - $ref: '#/components/parameters/dateFrom'
        - $ref: '#/components/parameters/dateTo'
        - name: amount_min
          in: query
          schema:
            type: number
            minimum: 0
        - name: amount_max
          in: query
          schema:
            type: number
            minimum: 0
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/offset'
        - name: sort
          in: query
          schema:
            type: string
            enum: [created_at, financial_loss, relevance]
            default: created_at
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ============ CASE UPDATE ============
...
  /admin/reports:
    get:
      tags: [Admin Reports]
      summary: List all reports (admin)
      description: Get paginated list of all reports with full details
      operationId: adminListReports
      security:
        - BearerAuth: [admin:read]
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
        - $ref: '#/components/parameters/status'
        - name: severity
          in: query
          schema:
            type: string
            enum: [LOW, MEDIUM, HIGH, CRITICAL]
        - $ref: '#/components/parameters/fraudType'
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, severity, status]
            default: createdAt
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Reports list with stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminReportListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: HttpOnly cookie authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Webhook secret for external services
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    status:
      name: status
      in: query
      schema:
        type: string
        enum: [pending, under_review, approved, rejected, merged, archived]
    fraudType:
      name: fraud_type
      in: query
      schema:
        type: string
        enum: [PHISHING, SCAM, FAKE_SHOP, INVESTMENT_FRAUD, ROMANCE_SCAM, TECH_SUPPORT, LOTTERY, CRYPTOCURRENCY, IDENTITY_THEFT, OTHER]
    country:
      name: country
      in: query
      schema:
        type: string
        maxLength: 2
      description: ISO 3166-1 alpha-2 country code
    dateFrom:
      name: date_from
      in: query
      schema:
        type: string
        format: date
    dateTo:
      name: date_to
      in: query
      schema:
        type: string
        format: date
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          description: Request ID for debugging
        timestamp:
          type: string
          format: date-time
      required: [error, message]
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            details:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
            issues:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    TokenResponse:
      type: object
      properties:
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token expiration in seconds
        scopes:
          type: array
          items:
            type: string
        user:
          $ref: '#/components/schemas/UserBasic'
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token (not included for API key auth)
        requires_2fa:
          type: boolean
          description: True if 2FA verification required
        temp_token:
          type: string
          description: Temporary token for 2FA verification
...
    UserBasic:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [BASIC, STANDARD, GOLD, ADMIN, SUPER_ADMIN]
...
    SearchResponse:
      type: object
      properties:
        total:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResult'
        facets:
          type: object
          properties:
            country:
              type: object
              additionalProperties:
                type: integer
            fraud_type:
              type: object
              additionalProperties:
                type: integer
        pagination:
          $ref: '#/components/schemas/Pagination'
    SearchResult:
      type: object
      properties:
        id:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
        source:
          type: string
          enum: [exact, fuzzy, semantic]
        perpetrator:
          type: object
          properties:
            name:
              type: string
              nullable: true
            phone:
              type: string
              nullable: true
            email:
              type: string
              nullable: true
        fraud_type:
          type: string
        country:
          type: string
        incident_date:
          type: string
          format: date
        highlights:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
        pages:
          type: integer
        total_pages:
          type: integer
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # ============ STATS SCHEMAS ============
...
    AdminReportListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/AdminReportSummary'
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
        stats:
          type: object
          properties:
            pending:
              type: integer
            underReview:
              type: integer
            approved:
              type: integer
            rejected:
              type: integer
            merged:
              type: integer
            archived:
              type: integer
    AdminReportSummary:
      type: object
      properties:
        id:
          type: string
        publicId:
          type: string
        caseNumber:
          type: string
        status:
          type: string
        severity:
          type: string
          nullable: true
        fraudType:
          type: string
        summary:
          type: string
        financialLoss:
          type: object
          nullable: true
          properties:
            amount:
              type: number
            currency:
              type: string
        country:
          type: string
        reporter:
          type: object
          nullable: true
          properties:
            id:
              type: string
            email:
              type: string
            displayName:
              type: string
        moderatedBy:
          type: object
          nullable: true
          properties:
            id:
              type: string
            displayName:
              type: string
        counts:
          type: object
          properties:
            perpetrators:
              type: integer
            evidence:
              type: integer
            comments:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
          nullable: true
        moderatedAt:
          type: string
          format: date-time
          nullable: true
...
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
            message: "Missing required scope: admin:read"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limited
            message: Rate limit exceeded. Try again later.